Added new migration; Added a color helper class; Added device repository; Added some changes to the main class; Added device repository interface;

+++ DeviceManagementSystem/Models/Device.cs
@@ -15,7 +15,7 @@
         {
         }
 
-        public int DeviceId { get; set; }
+        public long DeviceId { get; set; }
         public string Model { get; set; }
         public string Manufacturer { get; set; }
         public DateTime ManufacturDate { get; set; }
